handle correctly "anonymous" array/object (defect reported by QE)  (#137)
* a fix for defect reported by QE for JSON input.
the fix handle the use cases where the JSON input starts with an anonymous array/object
this may cause wrong search result per the user request(SQL statement)

* editorial

* adding test that verifies correct key-value extraction upon anonymous array

* adding relevant tests per changes.
handle the use-case where the user requests a json-key-path that may point to non-discrete value. i.e. array or an object.
editorial changes.

* a fix for CSV flow.
upon a broken row, the use of csv_parser->next_row() is wrong, since the row may break within a quote
and that could result with an exception.
the csv_parser should init with the csv-stream after skipping the bytes of the broken row

* editorial

* replace ifstream read API

* upon where-clause return empty group aggregation function returned the wrong result.
it should return a null

* add an operator to display the version description of current engine functionalities(recent PR)

* add tests per  aggregation function with an empty group

Signed-off-by: galsalomon66 <gal.salomon@gmail.com>